+++ tsup.config.ts
@@ -12,6 +12,7 @@ export default defineConfig({
     'src/components/data-table/index.ts',
     'src/components/layout/index.ts',
     'src/components/page/index.ts',
+    'src/components/theme/index.ts',
     'src/lib/utils.ts',
     'src/hooks/use-mobile.tsx',
     // Never build test files.